% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overwrite_library.R
\name{fix_library_calls}
\alias{fix_library_calls}
\title{Reformat from library calls}
\usage{
fix_library_calls(
  file,
  consider_loaded_namespaces = FALSE,
  remove_lib_calls = FALSE
)
}
\arguments{
\item{file}{The file to fix the calls in.}

\item{consider_loaded_namespaces}{Look for potential matches to function in loaded namespaces.}

\item{remove_lib_calls}{Remove the calls to library/require after replacements are done.}
}
\value{
The fixed code (as a string).
}
\description{
Replace calls to library with namespace::functions
}
